requirements.txt
asn1crypto==0.24.0
astroid==2.0.1
autopep8==1.3.5
beautifulsoup4==4.6.3
cachetools==2.1.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
coverage==4.5.1
coveralls==1.3.0
cryptography==2.3.1
defusedxml==0.5.0
Django==2.0.7
django-braces==1.13.0
django-cors-middleware==1.3.1
django-extensions==2.1.0
django-filter==2.0.0
django-oauth-toolkit==1.2.0
django-rest-framework-social-oauth2==1.1.0
django-sendgrid-v5==0.6.893
django-taggit==0.22.2
django-taggit-serializer==0.1.7
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
djmail==1.1.0
docopt==0.6.2
facebook-sdk==3.0.0
future==0.16.0
google-auth==1.5.1
gunicorn==19.9.0
idna==2.7
isort==4.3.4
lazy-object-proxy==1.3.1
Markdown==2.6.11
mccabe==0.6.1
MiniMock==1.2.8
oauthlib==2.1.0
Pillow==5.2.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycodestyle==2.4.0
pycparser==2.18
PyJWT==1.4.2
pylint==2.0.1
pyOpenSSL==18.0.0
python-http-client==3.1.0
python3-openid==3.1.0
pytz==2018.5
requests==2.19.1
requests-oauthlib==1.0.0
rsa==3.4.2
sendgrid==5.4.1
six==1.10.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
urllib3==1.23
whitenoise==3.3.1
wrapt==1.10.11
